Some changes with ctags.
Omni-complete for c.  For some reason you have to explicitly load
the auotoloads.  Which means I should add an autocmd line for
each of the languages I use. :(

Also, I built a HUGE tags database for the system include directories,
but omni-complete chokes hard on it, so I've told vim not to read the
database.